+++ cptkip/device/audio.py
@@ -1,9 +1,3 @@
-# See the following sources for reference:
-#  * https://docs.circuitpython.org/en/latest/shared-bindings/audiopwmio/
-#  * https://docs.circuitpython.org/en/latest/shared-bindings/audiomp3/
-#  * https://learn.adafruit.com/circuitpython-essentials/circuitpython-audio-out
-#  * https://learn.adafruit.com/circuitpython-essentials/circuitpython-mp3-audio
-#
 import cptkip.core.environment as environment
 
 if environment.are_pins_available():
@@ -13,36 +7,22 @@ if environment.are_pins_available():
     except ImportError:
         pass
 
-    try:
-        # noinspection PyPackageRequirements
-        from audioio import AudioOut
-    except ImportError:
-        try:
-            # noinspection PyUnresolvedReferences
-            from audiopwmio import PWMAudioOut as AudioOut
-        except ImportError:
-            pass  # not always supported by every board!
-
 
 class Audio:
     """
-    Audio wraps up AudioOut and an MP3 Decoder to make it simpler to play
-    music. It is a relatively light wrapper buts saves some boilerplate.
+    Audio wraps an MP3 Decoder to make it simpler to play music. It is a relatively
+    light wrapper buts saves some boilerplate. It requires an audio_out to play the
+    music such as an AudioOut or I2SOut.
     """
 
-    def __init__(self, pin):
-        """
-        :param pin: The pin to output audio on.
-        """
-        if environment.are_pins_available() and pin is None:
-            raise ValueError("pin cannot be None")
-
-        self.pin = pin
-        self.audio = None
+    def __init__(self, audio_out):
+        self.audio_out = audio_out
         self.decoder = None
 
-        if environment.are_pins_available():
-            self.audio = AudioOut(pin)
+        if environment.are_pins_available() and audio_out is None:
+            raise ValueError("audio cannot be None")
+
+        if audio_out is not None:
             # You have to specify some mp3 file when creating the decoder
             decoder = MP3Decoder(open("cptkip/mp3.mp3", "rb"))
             self.decoder = decoder
@@ -51,11 +31,11 @@ class Audio:
         """
         Releases the decoder and audio output. Safe to call multiple times.
         """
-        if self.audio:
+        if self.audio_out:
             self.decoder.deinit()
-            self.audio.deinit()
+            self.audio_out.deinit()
             self.decoder = None
-            self.audio = None
+            self.audio_out = None
             self.pin = None
 
     def play(self, filename: str) -> None:
@@ -67,44 +47,104 @@ class Audio:
         if filename is None or len(filename) <= 0:
             raise ValueError("filename must be specified")
 
-        if self.audio:
+        if self.audio_out:
             self.decoder.file = open(filename, "rb")
-            self.audio.play(self.decoder)
+            self.audio_out.play(self.decoder)
 
     @property
     def playing(self) -> bool:
         """
         Returns whether audio is currently playing.
         """
-        return self.audio.playing if self.audio else False
+        return self.audio_out.playing if self.audio_out else False
 
     @property
     def paused(self) -> bool:
         """
         Returns whether playback is currently paused.
         """
-        return self.audio.paused if self.audio else False
+        return self.audio_out.paused if self.audio_out else False
 
     def pause(self) -> None:
         """
         Pauses playback.
         """
-        if self.audio:
-            self.audio.pause()
+        if self.audio_out:
+            self.audio_out.pause()
 
     def resume(self) -> None:
         """
         Resumes playback after a pause.
         """
-        if self.audio:
-            self.audio.resume()
+        if self.audio_out:
+            self.audio_out.resume()
 
     def stop(self) -> None:
         """
         Stops playback.
         """
-        if self.audio:
-            self.audio.stop()
+        if self.audio_out:
+            self.audio_out.stop()
+
+
+# See the following sources for reference:
+#  * https://docs.circuitpython.org/en/latest/shared-bindings/audiopwmio/
+#  * https://docs.circuitpython.org/en/latest/shared-bindings/audiomp3/
+#  * https://learn.adafruit.com/circuitpython-essentials/circuitpython-audio-out
+#  * https://learn.adafruit.com/circuitpython-essentials/circuitpython-mp3-audio
+# noinspection pep8-naming
+def PwmAudio(pin) -> Audio:
+    """
+    Returns an Audio that uses a single pin to play music via PWM.
+
+    :param pin: The pin to output audio on.
+    """
+    if environment.are_pins_available() and pin is None:
+        raise ValueError("pin cannot be None")
+
+    audio_out = None
+
+    if environment.are_pins_available():
+        try:
+            # noinspection PyPackageRequirements
+            from audioio import AudioOut
+        except ImportError:
+            try:
+                # noinspection PyUnresolvedReferences
+                from audiopwmio import PWMAudioOut as AudioOut
+            except ImportError:
+                pass  # not always supported by every board!
+
+        audio_out = AudioOut(pin)
+
+    return Audio(audio_out)
+
+
+# See the following sources for reference:
+#  * https://docs.circuitpython.org/en/latest/shared-bindings/audiobusio/
+#  * https://learn.adafruit.com/mp3-playback-rp2040/pico-i2s-mp3
+#  * https://learn.adafruit.com/i2s-amplifier-bff/circuitpython
+# noinspection pep8-naming
+def I2sAudio(bit_clock, left_right_clock, data) -> Audio:
+    """
+    Returns an Audio that uses three pins to play music via PWM.
+    """
+    if environment.are_pins_available():
+        if bit_clock is None:
+            raise ValueError("bit_clock cannot be None")
+        if left_right_clock is None:
+            raise ValueError("left_right_clock cannot be None")
+        if data is None:
+            raise ValueError("data cannot be None")
+
+    audio_out = None
+
+    if environment.are_pins_available():
+        # noinspection unresolved-references
+        from audiobusio import I2SOut
+        audio_out = I2SOut(bit_clock, left_right_clock, data)
+
+    return Audio(audio_out)
 
 
 class Queue:
@@ -114,8 +154,6 @@ class Queue:
     played through the Audio instance. Basic controls to pause, resume and
     stop are provided along with a cancel option which stops the music and
     clears the queue.
-
-    Instances of this class will need to register() with a Runner in order to work.
     """
 
     def __init__(self, audio: Audio):
@@ -123,7 +161,7 @@ class Queue:
             raise ValueError("audio cannot be None")
 
         if not isinstance(audio, Audio):
-            raise ValueError("audio must be of type Audio")
+            raise ValueError("audio must be of type PwmAudio or I2SAudio")
 
         self._audio = audio
         self._queue = []