+++ examples/4 - Devices/g_i2s_audio.py
@@ -1,13 +1,13 @@
 #
-# This example plays an MP3 audio file using the PwmAudio and Queue classes.
+# This example plays an MP3 audio file using the I2SAudio and Queue classes.
 # A Button is used to add more "songs" to the queue as well as pause/unpause
-# the audio. This example is almost identical to g_i2sß_audio.py.
+# the audio. This example is almost identical to f_pwm_audio.py.
 #
 import time
 
 import cptkip.config.configuration as config
 import cptkip.core.logging as log
-from cptkip.device.audio import PwmAudio as Audio, Queue
+from cptkip.device.audio import I2sAudio as Audio, Queue
 from cptkip.device.button import Button
 from cptkip.pin.input_pin import InputPin
 
@@ -15,7 +15,7 @@ log.set_log_level(log.INFO)
 
 AUDIO_FILE = "examples/lion.mp3"
 
-audio = Audio(config.BUZZER_PIN)
+audio = Audio(config.I2S_BIT_CLOCK, config.I2S_LEFT_RIGHT_CLOCK, config.I2S_DATA)
 queue = Queue(audio)