+++ validate/device/i2s_audio.py
@@ -1,9 +1,9 @@
 def execute():
     import cptkip.config.configuration as config
-    from cptkip.device.audio import PwmAudio as Audio, Queue
+    from cptkip.device.audio import I2sAudio as Audio, Queue
     import validate.utils as utils
 
-    audio = Audio(config.BUZZER_PIN if config.BUZZER_PIN else "none")
+    audio = Audio(config.I2S_BIT_CLOCK, config.I2S_LEFT_RIGHT_CLOCK, config.I2S_DATA)
     queue = Queue(audio)
     queue.queue("validate/lion.mp3")
     queue.queue("validate/lion.mp3")
@@ -11,7 +11,7 @@ def execute():
     def task():
         queue.update()
 
-    print("Lion roar will play through buzzer")
+    print("Lion roar will play through speaker")
     utils.execute(task)
 
     audio.deinit()